Better OSC8 link handling.
Made osc8 links route through the wordstack, and made a pref for if links can
click directly, or have to open the wordstack first.  This is because moble
devices don't have a 'mouse hover' to be able to see what player is about to
click.  Has to click it first to see the preview in wordstack.